library/src/test/java/com/ibm/research/drl/dpt/providers/identifiers/EmailIdentifierTest.java
/*
Licensed to the Apache Software Foundation (ASF) under one
or more contributor license agreements.  See the NOTICE file
distributed with this work for additional information
regarding copyright ownership.  The ASF licenses this file
to you under the Apache License, Version 2.0 (the
"License"); you may not use this file except in compliance
with the License.  You may obtain a copy of the License at

  http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing,
software distributed under the License is distributed on an
"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
KIND, either express or implied.  See the License for the
specific language governing permissions and limitations
under the License.
*/
package com.ibm.research.drl.dpt.providers.identifiers;

import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 */
public class EmailIdentifierTest {

    @Test
    public void testIsOfThisType() throws Exception {
        EmailIdentifier identifier = new EmailIdentifier();

        assertTrue(identifier.isOfThisType("john@ie.ibm.com"));
        assertTrue(identifier.isOfThisType("santonat@ie.ibm.com"));
        assertFalse(identifier.isOfThisType("something else"));
        assertFalse(identifier.isOfThisType("Help@IBM"));
    }

    @Test
    @Disabled
    public void testQuickCheckBenefit() {
        EmailIdentifier identifier = new EmailIdentifier(); 
        int runs = 1000000;

        String value = "this is definitely not a match";

        long start = System.currentTimeMillis();
        int matches = 0;
        for(int i = 0; i < runs; i++) {
            if (identifier.isOfThisType(value)) {
                matches++;
            }
        }
        long end = System.currentTimeMillis();
        System.out.println("time : " + (end - start));
        System.out.println("matches: " + matches);

        start = System.currentTimeMillis();
        matches = 0;
        for(int i = 0; i < runs; i++) {
            if (quickCheck(value)) {
                if (identifier.isOfThisType(value)) {
                    matches++;
                }
            }
        }
        end = System.currentTimeMillis();
        System.out.println("time : " + (end - start));
        System.out.println("matches: " + matches);
    }

    private boolean quickCheck(String value) {
        return value.indexOf('@') != -1;
    }
}
